:building_construction: Restructure folders
