+++ src/util/images.ts
@@ -1,4 +1,4 @@
 declare module '*.svg' {
-  const content: any;
+  const content: string;
   export default content;
 }
